Renamed parser to formula_parser
